cli/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "portfolio-cli"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "httpx>=0.27",
    "rich>=13.7",
    "keyring>=25.2",
    "python-dotenv>=1.0",
]

[project.scripts]
portfolio = "portfolio_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]